refactor: 모듈화 및 코드 최적화
