+++ index.html
@@ -79,9 +79,9 @@
       </div>
       <button class="check" id="checkResult">결과 확인하기</button>
     </div>
-    <div class="copyright">
+    <footer class="copyright">
       <p>Copyright 2024. DongAncourse</p>
-    </div>
+    </footer>
     <div class="floating-content" id="floating-content">
       <div class="floating-content-close-container">
         <span id="close-floating-content" class="floating-content-close">&times;</span>
@@ -127,12 +127,12 @@
           </li>
         </ul>
       </div>
-      <div class="floating-content-rate" id="">
+      <div class="floating-content-rate">
         <span>당신의 총 수익률은 %입니다.</span>
       </div>
       <button class="floating-content-button" id="restart">다시 시작하기</button>
     </div>
     <div class="popup-background" id="popup-background"></div>
-    <script src="js/script.js"></script>
+    <script type="module" src="./js/script.js"></script>
   </body>
 </html>
